Update from original

+++ java/java-impl/src/com/intellij/codeInspection/java18api/Java8MapApiInspection.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2000-2016 JetBrains s.r.o.
+ * Copyright 2000-2017 JetBrains s.r.o.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -143,7 +143,7 @@ public class Java8MapApiInspection extends BaseJavaBatchLocalInspectionTool {
           condition.register(holder, new ReplaceWithSingleMapOperation("putIfAbsent", getCall, putArgs[1], result));
         }
         if (mySuggestMapGetOrDefault && condition.isContainsKey() && ExpressionUtils.isSimpleExpression(noneExpression) &&
-          !(getCall.getType() instanceof PsiCapturedWildcardType)) {
+          condition.isMapValueType(noneExpression.getType())) {
           condition.register(holder, new ReplaceWithSingleMapOperation("getOrDefault", getCall, noneExpression, result));
         }
       }
@@ -157,11 +157,10 @@ public class Java8MapApiInspection extends BaseJavaBatchLocalInspectionTool {
               value = ...
             }
            */
-          if (ExpressionUtils.isSimpleExpression(assignment.getRExpression()) &&
-              condition.isValueReference(assignment.getLExpression()) &&
-              !condition.isValueReference(assignment.getRExpression())) {
-            condition
-              .register(holder, ReplaceWithSingleMapOperation.fromIf("getOrDefault", condition, assignment.getRExpression()));
+          PsiExpression rValue = assignment.getRExpression();
+          if (ExpressionUtils.isSimpleExpression(rValue) && condition.isValueReference(assignment.getLExpression()) &&
+              !condition.isValueReference(rValue) && condition.isMapValueType(rValue.getType())) {
+            condition.register(holder, ReplaceWithSingleMapOperation.fromIf("getOrDefault", condition, rValue));
           }
         } else if (condition.isGetNull()) {
           /*
@@ -465,7 +464,7 @@ public class Java8MapApiInspection extends BaseJavaBatchLocalInspectionTool {
         ct.deleteAndRestoreComments(conditional);
       }
       PsiVariable variable = condition.extractDeclaration();
-      if(variable != null && ReferencesSearch.search(variable).findFirst() == null) {
+      if (variable != null && !PsiTreeUtil.isAncestor(result, variable, true) && ReferencesSearch.search(variable).findFirst() == null) {
         new CommentTracker().deleteAndRestoreComments(variable);
       }
       CodeStyleManager.getInstance(project).reformat(result);
@@ -559,6 +558,12 @@ public class Java8MapApiInspection extends BaseJavaBatchLocalInspectionTool {
     public void register(ProblemsHolder holder, ReplaceWithSingleMapOperation fix) {
       //noinspection DialogTitleCapitalization
       holder.registerProblem(getFullCondition(), QuickFixBundle.message("java.8.map.api.inspection.description", fix.myMethodName), fix);
+    }
+
+    public boolean isMapValueType(PsiType type) {
+      PsiType mapExpressionType = myMapExpression.getType();
+      PsiType valueTypeParameter = PsiUtil.substituteTypeParameter(mapExpressionType, CommonClassNames.JAVA_UTIL_MAP, 1, false);
+      return valueTypeParameter != null && valueTypeParameter.isAssignableFrom(type);
     }
   }
 }
